requirements.txt
numpy==1.18.5
requests==2.24.0
matplotlib==3.2.2
pandas==1.0.5
seaborn==0.10.1
Pillow==8.1.2
scikit_learn==0.24.1
